First pass at concatenating anchors inside blocks
Anchors nested inside a text block must be merged up into the parent
text block.

I had previously tried to do this in a second pass after a naive scan
would retrieve all blocks.

Unfortunately, it looks like storing the nodes in the blocks would lose
their references when the blocks were passed out of the scan function.
So I decided to to the concatenating while doing the scan (which may not
be that much more complicated anyways, and possibly faster?).

Basically, during the scan, for each anchor I:

- check if the containing node is a "text block" (has a tag that we use
to deliniate text blocks).
- check if there's a following text node. If yes, concatenate into the
anchor.
- check if there's previous text node. If yes, modify the block that's
on the stack of blocks to add on the info from the anchor block. Because
the following block has already been concatenated onto the anchor block,
all information from from the anchor and surrounding blocks is now in
the last block on the stack.
- if there's no previous text node, then that means that the anchor (and
possibly the following block that's been concatenated) needs to be
pushed onto the end of the stack of blocks.
- fin

Some changes to accomodate the above...
- TextBlock renamed to Block, now holds anchor word count instead of
anchor word density. Anchor density has to be calculated in a pass after
concatenation.